LayoutTests/media/media-controller-playbackrate-expected.txt

EVENT(canplay)
EVENT(play)
EVENT(timeupdate)
EXPECTED (video.currentTime != '0') OK
END OF TEST
